utils: Move repository buffer readers
Repository-backed buffer helpers lived under the editor package, which made
callers reach through the editing surface for Git object and working-tree
reads.

The helpers now live in utils.repository_buffers, and editor exports only
editing and line-ending APIs. Batch, data, and command callers import
repository readers from the data layer while editor-bound editing imports
stay in place.

This commit moves repository buffer loading to the data layer.

The package tree now separates text editing from repository I/O ownership.
