+++ src/git_stage_batch/batch/attribution.py
@@ -20,7 +20,7 @@ from ..core.line_selection import parse_line_selection
 from ..core.models import LineLevelChange
 from ..data.consumed_selections import read_consumed_file_metadata
 from ..core.buffer import buffer_matches
-from ..editor import (
+from ..utils.repository_buffers import (
     load_git_blob_as_buffer,
     load_git_object_as_buffer_or_empty,
     load_working_tree_file_as_buffer,
