+++ src/git_stage_batch/batch/ownership.py
@@ -17,12 +17,12 @@ from ..core.buffer import (
     LineBuffer,
     buffer_byte_chunks,
 )
-from ..editor import (
-    Editor,
+from ..utils.repository_buffers import (
     load_git_blob_as_buffer,
     load_git_object_as_buffer,
     load_working_tree_file_as_buffer,
 )
+from ..editor import Editor
 from ..exceptions import AtomicUnitError, MergeError
 from ..i18n import _
 from ..utils.git import (
