+++ src/git_stage_batch/batch/source_refresh.py
@@ -17,7 +17,10 @@ from ..data.batch_sources import (
     load_session_batch_sources,
     save_session_batch_sources,
 )
-from ..editor import load_git_object_as_buffer, load_working_tree_file_as_buffer
+from ..utils.repository_buffers import (
+    load_git_object_as_buffer,
+    load_working_tree_file_as_buffer,
+)
 from .lineage import BatchSourceLineage
 from .match import match_lines
 from .ownership import (
