+++ src/git_stage_batch/commands/include.py
@@ -102,7 +102,7 @@ from ..core.buffer import (
     LineBuffer,
     buffer_matches,
 )
-from ..editor import (
+from ..utils.repository_buffers import (
     load_git_object_as_buffer,
     load_working_tree_file_as_buffer,
 )
