+++ src/git_stage_batch/commands/include_from.py
@@ -56,7 +56,7 @@ from ..core.buffer import (
     LineBuffer,
     write_buffer_to_working_tree_path,
 )
-from ..editor import (
+from ..utils.repository_buffers import (
     load_git_object_as_buffer,
     load_working_tree_file_as_buffer,
 )
