+++ src/git_stage_batch/data/text_lifecycle_detection.py
@@ -6,7 +6,7 @@ from contextlib import ExitStack
 
 from ..core.buffer import LineBuffer, buffer_byte_count
 from ..core.text_lifecycle import TextFileChangeType
-from ..editor import load_git_object_as_buffer
+from ..utils.repository_buffers import load_git_object_as_buffer
 from ..utils.git import get_git_repository_root_path
